fix: Authentication failure during install
Due to reasons, the mirror URL is only propagated to the Authenticator
of the Repository class, created by this Plugin. All other instances
of Authenticator still fall back to empty user and password.

To fix this I've added the repository URL to the config, which
the Authenticator uses to determin credentials.

Signed-off-by: Philipp Krueger <philipp.krueger@cariad.technology>